// src/app/shared/enums/api.enum.ts
export enum ApiEnum {
  AUTH= '/auth',
  REFRESH_TOKEN= '/auth/refresh',
  EMAIL_EXISTS= '/auth/user-exists',
  USER_API = '/admin/login',
  ADMIN_PRODUCT = '/admin/product',
  PRODUCT = '/product',
  ADMIN_ORDER = '/admin/order',
  ORDER = '/order',
  ADMIN_CART = '/admin/cart',
  CART = '/cart',
  ROLE_API = '/admin/role',
  PERMISSION_API = '/admin/permission',
  PROFILE_INFORMATION = '/person',
  REGISTRATION = '/registration',
  ADMIN_REGISTRATION = '/admin/registration'
}